# src/types/Responses/waiter.oder.responses.ts
export interface SaveUpdateReswaiterOder {
    oder_id:string;
    Iteam: string;
    Desgination: string;
    cell: string;
    email: string;
    password:string;
    createdAt: string;
    updatedAt: string;
  
  }